// config/types.ts
export interface ChartDays {
  label: string;
  value: number;
}

export interface HistoricalChartData {
  prices: Array<number[]>;
  market_caps: Array<number[]>;
  total_volumes: Array<number[]>;
}